Fixed Mat::isContinuous for big matrix (thanks to Boris Mansencal) #2103

+++ modules/core/src/matrix.cpp
@@ -151,7 +151,7 @@ static void updateContinuityFlag(Mat& m)
     }
 
     int64 t = (int64)m.step[0]*m.size[0];
-    if( j <= i && t == (int)t )
+    if( j <= i && t == (size_t)t )
         m.flags |= Mat::CONTINUOUS_FLAG;
     else
         m.flags &= ~Mat::CONTINUOUS_FLAG;
